major restructure

+++ qctoolkit/QM/qmcode/vasp.py
@@ -1,18 +1,13 @@
 import qctoolkit as qtk
-import setting_pw as pw
-import pwinp as qin
+from qctoolkit.QM.pwinp import PlanewaveInput
 import os, sys, copy, shutil, re
 import numpy as np
-import qmjob
+import qctoolkit.QM.qmjob as qmjob
 import xml.etree.ElementTree as ET
 
-class inp(qin.PwInp):
-  def __init__(self, structure_inp, info, **kwargs):
-    qin.PwInp.__init__(self, structure_inp, info, **kwargs)
-    if 'PP' in kwargs:
-      self.PP = kwargs['PP']
-    else:
-      self.PP = qtk.PP
+class inp(PlanewaveInput):
+  def __init__(self, molecule, **kwargs):
+    PlanewaveInput.__init__(self, molecule, **kwargs)
 
   def run(self, name=None, **run_kw):
     if not name:
@@ -23,56 +18,24 @@ class inp(qin.PwInp):
     self.write(name)
     return qmjob.QMRun(name, 'vasp',**run_kw)
     
-
-  def write(self, *args, **kwargs):
-    new_structure = copy.deepcopy(self.structure)
-    if len(args) == 1:
-      name = args[0]
-      path = name + '/'
-      cwd = os.getcwd()
-      if os.path.exists(name):
-        if 'no_warning' in kwargs and kwargs['no_warning']:
+  def write(self, name=None):
+    molecule = copy.deepcopy(self.molecule)
+    self.cm_check(molecule)
+    cwd = os.getcwd()
+    if name:
+      if os.path.exists(name) and not qtk.setting.no_warning:
+        qtk.prompt(name + ' exists, overwrite?')
+        try:
           shutil.rmtree(name)
-          os.mkdir(name)
-          os.chdir(name)
-        else:
-          qtk.warning("inp.write: path " + name + " exist, "+\
-                      "nothing to be done")
-      else:
-        os.mkdir(name)
-        os.chdir(name)
-    else: 
-      cwd = '.'
-      name = ''
-      path = ''
+        except:
+          qtk.exit("can not remove folder: " + name)
+      os.mkdir(name)
+      os.chdir(name)
+
     incar   = open('INCAR',  'w') if name else sys.stdout
     kpoints = open('KPOINTS','w') if name else sys.stdout
     poscar  = open('POSCAR', 'w') if name else sys.stdout
     potcar  = open('POTCAR', 'w') if name else sys.stdout
-
-    if not self.setting.set_center and self.setting.set_margin:
-      if not self.setting.set_celldm:
-        Rt = np.transpose(new_structure.R)
-        for i in xrange(0, 3):
-          self.setting.celldm[i] = (max(Rt[i]) - min(Rt[i]))\
-                                 + 2 * self.setting.margin
-        new_center=[min(Rt[i])-self.setting.margin \
-          for i in (0,1,2)]
-        new_structure.center(new_center)
-      else:
-        qtk.exit("celldm and margin is not compatible, "+\
-                 "also check 'center', 'shift'")
-
-    elif self.setting.set_center and not self.setting.set_margin:
-      new_structure.center(self.setting.center)
- 
-    elif self.setting.set_center and self.setting.set_margin:
-      qtk.exit("ERROR from QM.vasp.py->inp.write: " +\
-               "center and margin " + \
-               "can NOT be set simultaneously.")
- 
-    if self.setting.set_shift:
-      new_structure.shift(self.setting.shift)
 
     # !!!!!!!!!!! TODO !!!!!!!!!!!
     # Center molecule
@@ -99,37 +62,37 @@ class inp(qin.PwInp):
 
     # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
     # construct to POSCAR and POTCAR data
-    new_structure.sort()
-    type_index = new_structure.index
-    type_list = new_structure.type_list
-    atom_list = self.atom_list
-    Z = new_structure.Z
+    molecule.sort()
+    type_index = molecule.index
+    type_list = molecule.type_list
+    Z = molecule.Z
+
+    if not qtk.setting.vasp_pp:
+      qtk.exit("vasp pp_path not set")
+    else:
+      self.pp_path = qtk.setting.vasp_pp
+
     for atom_type in xrange(0,len(type_index)-1):
       type_n = type_index[atom_type+1] - type_index[atom_type]
       # check special PP folder
-      if atom_list.has_key(str(Z[type_index[atom_type]])):
-        key = str(Z[type_index[atom_type]])
-        AtomPP = os.path.join(self.PP, atom_list[key], 'POTCAR')
-        catPOTCAR(AtomPP)
-        del atom_list[str(Z[type_index[atom_type]])]
+        # not yet implemented
       # default PP path
-      else:
-        type_name = qtk.Z2n(Z[type_index[atom_type]])
-        AtomPP = os.path.join(self.PP, type_name, 'POTCAR')
-        catPOTCAR(AtomPP)
+      type_name = type_list[type_index[atom_type]]
+      AtomPP = os.path.join(self.pp_path, type_name, 'POTCAR')
+      catPOTCAR(AtomPP)
       getNlist(type_n)
       for I in\
         xrange(type_index[atom_type],type_index[atom_type+1]):
-        getRlist(new_structure.R[I][:])
+        getRlist(molecule.R[I][:])
 
     # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
     # write to INCAR and generate POTCAR
-    qtk.report("vasp.inp", "writing", path+"INCAR")
-    print >> incar, "SYSTEM = %s" % self.info
+    qtk.report("vasp.inp", "writing", "INCAR")
+    print >> incar, "SYSTEM = %s" % self.setting['info']
     print >> incar, "ISMEAR = 0"
     print >> incar, "IBRION = 2"
-    if self.setting.set_vdw:
-      vdw = self.setting.vdw.lower()
+    if 'vdw' in self.setting:
+      vdw = self.setting['vdw'].lower()
       if vdw != 'none':
         if vdw=='d2':
           print >> incar, "IVDW = 10"
@@ -143,20 +106,21 @@ class inp(qin.PwInp):
           print >> incar, "IVDW = 212"
         else:
           qtk.exit("VDW '%s' is not supported for VASP" % vdw)
-    if new_structure.charge:
-      nve = new_structure.getValenceElectrons()
+    if molecule.charge:
+      nve = molecule.getValenceElectrons()
       print >> incar, "NELECT = %d" % (nve)
-    if not self.setting.save_density:
+    if 'save_density'not in self.setting\
+    or self.setting['save_density']:
       print >> incar, "LCHARG = .FALSE."
-    if 'scalapack' not in self.setting.settings:
+    if 'scalapack' not in self.setting:
       print >> incar, "LSCALAPACK = .FALSE."
-    elif not self.setting.settings['scalapack']:
+    elif not self.setting['scalapack']:
       print >> incar, "LSCALAPACK = .FALSE."
 
     # !!!!!!!!!!!!!!!!
     # write to KPOINTS
-    qtk.report("vasp.inp", "writin", path+"KPOINTS")
-    if not self.setting.kpoints:
+    qtk.report("vasp.inp", "writin", "KPOINTS")
+    if 'kmesh' not in self.setting:
       print >> kpoints, "Gamma-point only"
       print >> kpoints, " 1       ! one k-point"
       print >> kpoints, "rec      ! in units of reciprocal vector"
@@ -164,13 +128,13 @@ class inp(qin.PwInp):
 
     # !!!!!!!!!!!!!!!
     # write to POSCAR
-    qtk.report("vasp.inp", "writing", path+"POSCAR")
-    print >> poscar, self.info
+    qtk.report("vasp.inp", "writing", "POSCAR")
+    print >> poscar, self.setting['info']
     print >> poscar, "1.0"
-    self.setting.celldm2lattice()
+    self.celldm2lattice()
     for i in range(3):
       for j in range(3):
-        print >> poscar, " %7.4f" % self.setting.lattice[i,j],
+        print >> poscar, " %7.4f" % self.setting['lattice'][i,j],
       print >> poscar, "! lattic vector a(%d)" %i
     for n in n_list:
       print >> poscar, n,
@@ -183,7 +147,7 @@ class inp(qin.PwInp):
 
     # !!!!!!!!!!!!!!!
     # write to POTCAR
-    qtk.report("vasp.inp", "writing", path+"POTCAR")
+    qtk.report("vasp.inp", "writing", "POTCAR")
     for PP_file in PPPath:
       qtk.report("vasp.inp.POTCAR: ", PP_file)
       if name:
@@ -198,9 +162,7 @@ class inp(qin.PwInp):
     poscar.close()
     potcar.close()
 
-
-    if name:
-      os.chdir(cwd)
+    os.chdir(cwd)
 
 class out(object):
   """
